requirements.txt
tqdm
tensorflow
numpy
Pillow
pyperclip
pygame
mlib-nsi